+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### HEAD
 
+* Support renamed tasks
 * Prevent from releasing empty version numbers
 * bump multiple files at once (fixes geddski/grunt-release#90)